+++ packages/website/components/Layout/index.tsx
@@ -33,11 +33,12 @@ export default function (props: {
     if (!current.hasInit && !localStorage.getItem("saidHello")) {
       current.hasInit = true;
       localStorage.setItem("saidHello", "true");
-      console.log("🚀欢迎使用 VanBlog 博客系统");
-      console.log("当前版本：", props?.option?.version || "未知");
-      console.log("项目主页：", "https://vanblog.mereith.com");
-      console.log("开源地址：", "https://github.com/mereithhh/van-blog");
-      console.log("喜欢的话可以给个 star 哦🙏");
+      console.log('✨ Welcome to VanBlog Website ✨');
+      console.log('Version:', props?.option?.version || "unknown");
+      // console.log('Website:', 'https://vanblog.mereith.com');
+      console.log('GitHub:', 'https://github.com/CornWorld/vanblog');
+      console.log('!!!', 'This is a fork of VanBlog, and is not the official website.', '!!!');
+      console.log('If you like this project, please give it a star! 🌟');
       window.onbeforeunload = handleClose;
     }
     return () => {